requirements.txt
twikit==1.6.2
pandas==2.2.0
numpy==1.26.3
schedule==1.2.1
python-dotenv==1.0.1